[Refactor] 로또 관련 상수 Enterprise로 이동
