--- src/main/kotlin/constant/Constant.kt
@@ -1,9 +1,0 @@
-package constant
-
-const val LOTTO_NUM_COUNT = 6
-const val LOTTO_BONUS_NUM_COUNT = 1
-const val LOTTO_PRICE = 1_000
-const val LOTTO_NUM_RANGE_START = 1
-const val LOTTO_NUM_RANGE_END = 45
-const val LOTTO_MAX_GRADE = 5
-const val LOTTO_MIN_GRADE = 1
