+++ src/main/kotlin/winlotto/WinLotto.kt
@@ -1,8 +1,8 @@
 package winlotto
 
-import constant.LOTTO_BONUS_NUM_COUNT
-import constant.LOTTO_NUM_COUNT
-import constant.LOTTO_NUM_RANGE_END
+import enterprise.LOTTO_BONUS_NUM_COUNT
+import enterprise.LOTTO_NUM_COUNT
+import enterprise.LOTTO_NUM_RANGE_END
 import exception.DuplicateInputException
 import util.ErrorType
 import util.printErrorMessage
